// src/svelte/store/position/transform/types-local.d.ts
/**
 * Defines optional validation data for adjustment parameters.
 */
type ValidationData = {
   width?: number | null | undefined,
   marginTop?: number | null | undefined,
   offsetTop?: number | null | undefined,

   height?: number | null | undefined,
   marginLeft?: number | null | undefined,
   offsetLeft?: number | null | undefined
}

export { ValidationData }